Chore(Api): routes /v1/... -> /api/v1 and plural
/v1/culture/<name> is now /api/v1/cultures/<name> these changes have
been made for both the frontend api and the api itself. This follows
more of RESTful naming conventions and allows for us to host other pages
on this Api to display actual content like a Terms of Service. This is
how many sites are structured they have their normal routes that
render/display pages off of index and then their api off of /api/vXXX.